Added authorization

+++ backend/routes/posts.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const multer = require("multer");
+const checkAuth = require("../middleware/check-auth");
 const Post = require('../models/post');
 
 const router = express.Router();
@@ -25,12 +26,13 @@ const storage = multer.diskStorage({
     }
 });
 
-router.post("", multer({storage: storage}).single("image"), (req, res, next) => {
+router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
     const url = req.protocol + '://' + req.get('host');
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -43,7 +45,7 @@ router.post("", multer({storage: storage}).single("image"), (req, res, next) => 
     });    
 });
 
-router.put('/:id', multer({storage: storage}).single("image"), (req, res, next) => {
+router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
     let imagePath = req.body.imagePath;
     if (req.file) {
         const url = req.protocol + '://' + req.get('host');
@@ -53,15 +55,21 @@ router.put('/:id', multer({storage: storage}).single("image"), (req, res, next) 
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    Post.updateOne({_id: req.params.id}, post)
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post)
         .then(result => {
-            console.log(result);
-            res.status(200).json({
-                message: "Update was successful.",
-                imagePath: imagePath
-            });
+            if (result.modifiedCount > 0) {
+                res.status(200).json({
+                    message: "Update was successful."
+                });
+            }
+            else {
+                res.status(401).json({
+                    message: "User is not authorized."
+                });
+            }
         });
 });
 
@@ -102,12 +110,21 @@ router.get('', (req, res, next) => {
         });
 });
 
-router.delete('/:id', (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id })
+router.delete('/:id', checkAuth, (req, res, next) => {
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
         .then(result => {
             console.log(result);
-        })
-    res.status(200).json({message: 'Post deleted'});
+            if (result.deletedCount > 0) {
+                res.status(200).json({
+                    message: "Delete was successful."
+                });
+            }
+            else {
+                res.status(401).json({
+                    message: "User is not authorized."
+                });
+            }
+        });
 });
 
 module.exports = router;
